AB#4497 Added ProForma add command that calls the respective api call
This function takes a file path and sends the path and data to the api where it is attached to a proforma version

This will create a new pfv if the pfv has a file attached already if a file is not given it will pull the previous file instead

and if no file and a id is given it pull the file at a specific pfv version.

this commit only adds the attach functionality however this shed light on what type of response needs to be sent and verified.