AB#4710 Completed the flow of data for downloading
First we will get the sequences from the user then get the paths that will hit the api then we will got through all the paths to hit the api and fetch the download data. I'm not sure if the hash stuff is happening but I hope. After we have all the data we filter it further to remove unwanted things like inactive samples. All this api data then is used to hit the api again to initiate a download for all the files which go to a user specified output file.