AB#4822 Removed the ability to update a projects abbreviations.
This was done as the underlying group name does not change with it-as it is linked.

Since the link between group name and abbrev will be severed in the future. Removal seemed like the best course of action as it is a key system field that should not be changed.