AB#7217 fixed update project not updating merge algo
This commit addresses the removal of the redundant ⁠`set-merge-algorithm` functionality, which is no longer necessary. While this solution is the fastest way to resolve the current issue, it also entails additional work to eliminate the project settings table from the server.
In addition to removing the now dead endpoint `set merge algorithm`, we will also need to eliminate the ⁠set merge endpoint on the backend. This task can be combined with the effort to remove the project settings table.